convert double async id to int on syscall

+++ src/handle_wrap.h
@@ -87,7 +87,7 @@ class HandleWrap : public AsyncWrap {
                   double execution_async_id = kInvalidAsyncId,
                   bool silent = false) override {
     AsyncWrap::AsyncReset(resource, execution_async_id, silent);
-    syscall(516, get_async_id(), handle_);
+    syscall(516, static_cast<int64_t>(get_async_id()), handle_);
   }
 
  protected:
